Legg til inntektskontroll-tab i aktivitetspenger beregning
