+++ packages/v2/gui/src/prosess/aktivitetspenger-beregning/AktivitetspengerBeregning.tsx
@@ -1,133 +1,108 @@
+import {
+  ung_kodeverk_behandling_aksjonspunkt_AksjonspunktStatus as AksjonspunktDtoStatus,
+  type ung_sak_kontrakt_aksjonspunkt_AksjonspunktDto as AksjonspunktDto,
+  type ung_sak_kontrakt_kontroll_KontrollerInntektDto as KontrollerInntektDto,
+} from '@k9-sak-web/backend/ungsak/generated/types.js';
 import type { BeregningsgrunnlagDto } from '@k9-sak-web/backend/ungsak/kontrakt/aktivitetspenger/BeregningsgrunnlagDto.js';
-import { BesteBeregningResultatType } from '@k9-sak-web/backend/ungsak/kontrakt/aktivitetspenger/BesteBeregningResultatType.js';
-import { formatDate } from '@k9-sak-web/gui/utils/formatters.js';
-import { CheckmarkHeavyIcon } from '@navikt/aksel-icons';
-import { BodyShort, Box, Heading, HStack, InlineMessage, Table, VStack } from '@navikt/ds-react';
-import styles from './aktivitetspengerBeregning.module.css';
+import { aksjonspunktCodes } from '@k9-sak-web/backend/ungsak/kodeverk/AksjonspunktCodes.js';
+import { ArbeidOgInntekt } from '../../shared/kontroll-inntekt/ArbeidOgInntekt';
+import { ExclamationmarkTriangleFillIcon } from '@navikt/aksel-icons';
+import { Alert, Box, Heading, Loader, Tabs } from '@navikt/ds-react';
+import { useQuery } from '@tanstack/react-query';
+import type { AktivitetspengerBeregningBackendApiType } from './AktivitetspengerBeregningBackendApiType';
+import AktivitetspengerBeregningsgrunnlag from './AktivitetspengerBeregningsgrunnlag';
 
-const formatter = new Intl.NumberFormat('nb-NO', {
-  minimumFractionDigits: 0,
-  maximumFractionDigits: 2,
-});
-
-const SelectedCell = ({ children }: { children: string }) => (
-  <HStack gap="space-4" className={styles.selectedCell}>
-    <CheckmarkHeavyIcon title="Besteberegning" fontSize="1.5rem" />
-    <BodyShort>{children}</BodyShort>
-  </HStack>
-);
+const sortInntekt = (data: KontrollerInntektDto): KontrollerInntektDto => {
+  const { kontrollperioder } = data;
+  return {
+    kontrollperioder: kontrollperioder
+      ?.toSorted((a, b) => {
+        if (!a.periode || !b.periode) {
+          return 0;
+        }
+        return new Date(a.periode.fom).getTime() - new Date(b.periode.fom).getTime();
+      })
+      .toReversed(),
+  };
+};
 
 interface Props {
   data: BeregningsgrunnlagDto;
+  behandling: { uuid: string; versjon: number };
+  api: AktivitetspengerBeregningBackendApiType;
+  aksjonspunkter: AksjonspunktDto[];
+  submitCallback: (data: unknown) => Promise<any>;
+  isReadOnly: boolean;
 }
 
-const AktivitetspengerBeregning = ({ data }: Props) => {
-  const årstallForSkjæringstidspunkt = new Date(data.skjæringstidspunkt).getFullYear();
-  const isBesteberegningSnittSisteTreÅr =
-    data.besteBeregningResultatType === BesteBeregningResultatType.SNITT_SISTE_TRE_ÅR;
-  const isBesteberegningSisteÅr = data.besteBeregningResultatType === BesteBeregningResultatType.SISTE_ÅR;
-  const sortertePgiÅrsinntekter = data.pgiÅrsinntekter.toSorted((a, b) => b.årstall - a.årstall);
+const AktivitetspengerBeregning = ({ data, behandling, api, aksjonspunkter, submitCallback, isReadOnly }: Props) => {
+  const {
+    data: inntekt,
+    isLoading: kontrollInntektIsLoading,
+    isError: kontrollInntektIsError,
+  } = useQuery({
+    queryKey: ['aktivitetspenger-kontrollInntekt', behandling.uuid, behandling.versjon],
+    queryFn: () => api.getKontrollerInntekt(behandling.uuid),
+    select: sortInntekt,
+  });
+
+  const { data: arbeidsgivere } = useQuery({
+    queryKey: ['aktivitetspenger-arbeidsgivere', behandling.uuid],
+    queryFn: () => api.getArbeidsgiverOpplysninger(behandling.uuid),
+  });
+
+  if (kontrollInntektIsError) {
+    return <Alert variant="error">Noe gikk galt, vennligst prøv igjen senere</Alert>;
+  }
+
+  const harInntektKontroll = inntekt?.kontrollperioder && inntekt.kontrollperioder.length > 0;
+  const aksjonspunkt = aksjonspunkter?.find(ap => ap.definisjon === aksjonspunktCodes.KONTROLLER_INNTEKT);
+  const harUløstAksjonspunkt = aksjonspunkt && aksjonspunkt.status === AksjonspunktDtoStatus.OPPRETTET;
+  const defaultTab = aksjonspunkt && harInntektKontroll ? 'inntekt' : 'beregningsgrunnlag';
 
   return (
-    <VStack gap="space-8">
-      <Heading size="medium" level="1" spacing>
-        Grunnlag for beregning
-      </Heading>
-      <VStack gap="space-36">
-        <Box padding="space-16" background="info-soft" width="fit-content">
-          <InlineMessage size="small" status="info">
-            Skjæringstidspunkt for beregning {formatDate(data.skjæringstidspunkt)}
-          </InlineMessage>
-        </Box>
-        <div>
-          <Heading size="small" level="2" spacing>
-            Pensjonsgivende inntekt siste 3 år
-          </Heading>
-          {sortertePgiÅrsinntekter.length > 0 && (
-            <Box
-              marginBlock="space-8 space-0"
-              borderRadius="8"
-              borderWidth="1"
-              width="800px"
-              borderColor="neutral-subtle"
-            >
-              <Table>
-                <Table.Header>
-                  <Table.Row className={styles.rowWithSpacing}>
-                    <Table.HeaderCell>År</Table.HeaderCell>
-                    <Table.HeaderCell align="right">Arbeid/Frilans</Table.HeaderCell>
-                    <Table.HeaderCell align="right">Næring</Table.HeaderCell>
-                    <Table.HeaderCell align="right">Sum</Table.HeaderCell>
-                    <Table.HeaderCell align="right">Sum avkortet 6G</Table.HeaderCell>
-                    <Table.HeaderCell align="right">G-justert({årstallForSkjæringstidspunkt})</Table.HeaderCell>
-                  </Table.Row>
-                </Table.Header>
-                <Table.Body>
-                  {sortertePgiÅrsinntekter.map((pgi, index) => {
-                    const isBesteberegning = isBesteberegningSisteÅr && index === 0;
-                    return (
-                      <Table.Row key={pgi.årstall} className={styles.rowWithSpacing}>
-                        <Table.DataCell>{pgi.årstall}</Table.DataCell>
-                        <Table.DataCell align="right">{formatter.format(pgi.arbeidsinntekt)}</Table.DataCell>
-                        <Table.DataCell align="right">{formatter.format(pgi.næring)}</Table.DataCell>
-                        <Table.DataCell align="right">{formatter.format(pgi.sum)}</Table.DataCell>
-                        <Table.DataCell align="right">{formatter.format(pgi.sumAvkortet)}</Table.DataCell>
-                        <Table.DataCell align="right">
-                          {isBesteberegning ? (
-                            <SelectedCell>{formatter.format(pgi.sumAvkortetOgOppjustert)}</SelectedCell>
-                          ) : (
-                            formatter.format(pgi.sumAvkortetOgOppjustert)
-                          )}
-                        </Table.DataCell>
-                      </Table.Row>
-                    );
-                  })}
-                  <Table.Row className={`${styles.bottomCell} ${styles.rowWithSpacing}`}>
-                    <Table.HeaderCell scope="row" colSpan={5}>
-                      Gjennomsnittlig pensjonsgivende inntekt siste 3 år{' '}
-                    </Table.HeaderCell>
-                    <Table.DataCell align="right">
-                      {isBesteberegningSnittSisteTreÅr ? (
-                        <SelectedCell>{formatter.format(data.årsinntektSisteTreÅr)}</SelectedCell>
-                      ) : (
-                        formatter.format(data.årsinntektSisteTreÅr)
-                      )}
-                    </Table.DataCell>
-                  </Table.Row>
-                </Table.Body>
-              </Table>
-            </Box>
-          )}
-        </div>
-        <div>
-          <Heading size="small" level="2" spacing>
-            Beregning av dagsats
-          </Heading>
-          <Box
-            marginBlock="space-8 space-0"
-            borderRadius="8"
-            borderWidth="1"
-            width="800px"
-            borderColor="neutral-subtle"
-          >
-            <Table>
-              <Table.Body>
-                <Table.Row className={styles.rowWithSpacing}>
-                  <Table.DataCell scope="row">Beregningsgrunnlag redusert til 66 %</Table.DataCell>
-                  <Table.DataCell align="right">{formatter.format(data.beregningsgrunnlagRedusert)}</Table.DataCell>
-                </Table.Row>
-                {data.dagsats && (
-                  <Table.Row className={`${styles.bottomCell} ${styles.rowWithSpacing}`}>
-                    <Table.HeaderCell scope="row">Dagsats (beregningsgrunnlag/260 dager)</Table.HeaderCell>
-                    <Table.DataCell align="right">{formatter.format(data.dagsats)}</Table.DataCell>
-                  </Table.Row>
-                )}
-              </Table.Body>
-            </Table>
-          </Box>
-        </div>
-      </VStack>
-    </VStack>
+    <Box paddingInline="space-16 space-32" paddingBlock="space-8">
+      <Box minHeight="100svh">
+        <Heading size="medium" level="1" spacing>
+          Grunnlag for beregning
+        </Heading>
+        {kontrollInntektIsLoading ? (
+          <Loader size="large" />
+        ) : (
+          <Tabs defaultValue={defaultTab}>
+            <Tabs.List>
+              <Tabs.Tab value="beregningsgrunnlag" label="Beregningsgrunnlag" />
+              {harInntektKontroll && (
+                <Tabs.Tab
+                  value="inntekt"
+                  label="Inntektskontroll"
+                  icon={
+                    harUløstAksjonspunkt && (
+                      <ExclamationmarkTriangleFillIcon fontSize="1.5rem" color="var(--ax-text-warning-decoration)" />
+                    )
+                  }
+                />
+              )}
+            </Tabs.List>
+            <Tabs.Panel value="beregningsgrunnlag">
+              <AktivitetspengerBeregningsgrunnlag data={data} />
+            </Tabs.Panel>
+            {harInntektKontroll && (
+              <Box maxWidth="860px">
+                <Tabs.Panel value="inntekt">
+                  <ArbeidOgInntekt
+                    submitCallback={submitCallback}
+                    inntektKontrollperioder={inntekt?.kontrollperioder ?? []}
+                    isReadOnly={isReadOnly}
+                    arbeidsgivere={arbeidsgivere}
+                  />
+                </Tabs.Panel>
+              </Box>
+            )}
+          </Tabs>
+        )}
+      </Box>
+    </Box>
   );
 };
