+++ packages/v2/gui/src/shared/kontroll-inntekt/aksjonspunktArbeidOgInntekt.module.css
@@ -1,0 +1,3 @@
+.personIcon {
+  flex-shrink: 0;
+}
