+++ packages/v2/gui/src/shared/kontroll-inntekt/aksjonspunktArbeidOgInntekt.module.d.css.ts
@@ -1,0 +1,4 @@
+declare const styles: {
+  readonly "personIcon": string;
+};
+export = styles;
