+++ packages/v2/gui/src/shared/kontroll-inntekt/arbeidOgInntekt.module.d.css.ts
@@ -1,0 +1,8 @@
+declare const styles: {
+  readonly "checkmarkIcon": string;
+  readonly "exclamationmarkIcon": string;
+  readonly "firstDataCell": string;
+  readonly "firstHeaderCell": string;
+  readonly "lastRow": string;
+};
+export = styles;
