+++ packages/v2/gui/src/shared/kontroll-inntekt/detaljerOmInntekt.module.d.css.ts
@@ -1,0 +1,5 @@
+declare const styles: {
+  readonly "sum": string;
+  readonly "sumLabel": string;
+};
+export = styles;
